Use 32-bit for ids again
This needs to be fixed but breaks some code otherwise

+++ api/include/pd_menu.h
@@ -12,7 +12,7 @@ extern "C" {
 typedef struct PDMenuItem
 {
     const char* name;
-    unsigned short id;
+    unsigned int id;
     int key;
     int macMod;
     int winMod;
